+++ src/hooks/LevelTools.cpp
@@ -3,7 +3,7 @@
 
 using namespace geode::prelude;
 
-class $modify(OdysseyLevelTools, LevelTools)
+class $modify(LevelTools)
 {
 	static bool verifyLevelIntegrity(std::string verifyString, int levelID)
 	{
@@ -243,7 +243,7 @@ class $modify(OdysseyLevelTools, LevelTools)
 			level->m_levelString = LocalLevelManager::sharedState()->getMainLevelString(levelID);
 
 		level->m_levelID = levelID;
-		level->m_levelType = GJLevelType::Local;
+		level->m_levelType = GJLevelType::Main;
 
 		return level;
 	};